do only a opengl context change in UnrimpNode::update() when something has changed:  - first init is needed  - FBO is dirty (e.g. size of the node has changed)  - the example has changed
Bugfix: don't call QSGNode::markDirty() in UnrimpNode::setExample(). This method is called from the Gui Thread.
In Qt 5.1.x the SceneGraph renders under linux in a separate thread (before it was single threaded).
-> the rendering stopped when QSGNode::markDirty() was called in UnrimpNode::setExample()

Changes needed for Qt5.1.x:
On Qt5.1.0 QSGNode::preprocess() is only called when the Node should render its content.
(In Qt 5.0.x this method was always called when QQuickItem::update() was called.)
It seems that in Qt 5.1.0 the content of a QSGNode is chached(?) and the node only must rerender its content when the cache gets invalid/dirty (e.g. the material of the node is marked as dirty)
-> QSGNode::preprocess() is called when something in the QML scene changes (e.g. an animation is running)
or the node is explicitly marked as "dirty".
=> to have the same behaviour as in 5.0.x the Material of the Node is always marked as Dirty when UnrimpNode::update() is called
TODO: It would be better to do this only when a real change occured or the renderer should be called periodically (e.g. the InstancedQube example)